Moved rebuilder tests from AbstractExtendedPdoTest to a new test class RebuilderTest
